ref: move ReentrantCallHandler to shared utils
