+++ contracts/src/utils/mod.rs
@@ -5,6 +5,7 @@ pub mod math;
 pub mod metadata;
 pub mod nonces;
 pub mod pausable;
+pub mod reentrant_call_handler;
 pub mod structs;
 
 pub use metadata::Metadata;
